Introduce `BackendState trait`
