+++ core-backend/common/src/lib.rs
@@ -147,6 +147,44 @@ pub trait BackendExtError: Clone + Sized {
     fn into_termination_reason(self) -> TerminationReason;
 }
 
+pub trait BackendState {
+    type CoreError: Display;
+
+    fn err_mut(&mut self) -> &mut SyscallFuncError<Self::CoreError>;
+}
+
+pub trait IntoExtErrorForResult<T, Err, State>
+where
+    Err: Display,
+{
+    fn into_ext_error(
+        self,
+        state: &mut State,
+    ) -> Result<Result<T, u32>, ActorSyscallFuncError<Err>>;
+}
+
+impl<T, Err, State> IntoExtErrorForResult<T, Err, State> for Result<T, Err>
+where
+    Err: BackendExtError + Display,
+    State: BackendState<CoreError = Err>,
+{
+    fn into_ext_error(
+        self,
+        state: &mut State,
+    ) -> Result<Result<T, u32>, ActorSyscallFuncError<Err>> {
+        match self {
+            Ok(value) => Ok(Ok(value)),
+            Err(err) => {
+                *state.err_mut() = ActorSyscallFuncError::Core(err.clone()).into();
+                match err.into_ext_error() {
+                    Ok(ext_err) => Ok(Err(ext_err.encoded_size() as u32)),
+                    Err(err) => Err(ActorSyscallFuncError::Core(err)),
+                }
+            }
+        }
+    }
+}
+
 pub trait GetGasAmount {
     fn gas_amount(&self) -> GasAmount;
 }
