+++ core-backend/sandbox/src/funcs.rs
@@ -20,10 +20,11 @@ use crate::runtime::Runtime;
 use alloc::{format, string::String};
 use blake2_rfc::blake2b::blake2b;
 use codec::Encode;
-use core::{convert::TryInto, fmt::Display, marker::PhantomData};
+use core::{convert::TryInto, marker::PhantomData};
 use gear_backend_common::{
     memory::{MemoryAccessRecorder, MemoryOwner},
-    ActorSyscallFuncError, BackendExt, BackendExtError, SyscallFuncError, TerminationReason,
+    ActorSyscallFuncError, BackendExt, BackendExtError, IntoExtErrorForResult, SyscallFuncError,
+    TerminationReason,
 };
 use gear_core::{
     env::Ext,
@@ -40,39 +41,6 @@ use sp_sandbox::{HostError, ReturnValue, Value};
 const PTR_SPECIAL: u32 = i32::MAX as u32;
 
 pub(crate) type SyscallOutput = Result<ReturnValue, HostError>;
-
-trait IntoExtErrorForResult<T, Err, Ext>
-where
-    Err: Display,
-    Ext: gear_core::env::Ext,
-{
-    fn into_ext_error(
-        self,
-        state: &mut Runtime<Ext>,
-    ) -> Result<Result<T, u32>, ActorSyscallFuncError<Err>>;
-}
-
-impl<T, Err, Ext> IntoExtErrorForResult<T, Err, Ext> for Result<T, Err>
-where
-    Err: BackendExtError + Display + Clone,
-    Ext: gear_core::env::Ext<Error = Err>,
-{
-    fn into_ext_error(
-        self,
-        state: &mut Runtime<Ext>,
-    ) -> Result<Result<T, u32>, ActorSyscallFuncError<Err>> {
-        match self {
-            Ok(value) => Ok(Ok(value)),
-            Err(err) => {
-                state.err = ActorSyscallFuncError::Core(err.clone()).into();
-                match err.into_ext_error() {
-                    Ok(ext_err) => Ok(Err(ext_err.encoded_size() as u32)),
-                    Err(err) => Err(ActorSyscallFuncError::Core(err)),
-                }
-            }
-        }
-    }
-}
 
 pub(crate) struct FuncsHandler<E: Ext + 'static> {
     _phantom: PhantomData<E>,
